Refactor: se Agrego el usuario Alejandro al arreglo

+++ database/migrations/2024_08_28_004129_add_role_id_to_users_table.php
@@ -9,7 +9,9 @@ class AddRoleIdToUsersTable extends Migration
     public function up()
     {
         Schema::table('users', function (Blueprint $table) {
+            if (!Schema::hasColumn('users', 'role_id')) {
             $table->foreignId('role_id')->constrained('roles')->onDelete('cascade');
+            }
         });
     }
